Se  crea el encarpeta base de la app
Ahora ya están los encarpetados correctos del proyecto
