+++ src/app/app.module.ts
@@ -3,7 +3,7 @@ import { BrowserModule } from '@angular/platform-browser';
 
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
-import { ButtonComponent } from './atoms/button/button.component';
+import { ButtonComponent } from './components/atoms/button/button.component';
 
 @NgModule({
   declarations: [
